refactor(pipeline-rdf): kebab-case module filenames (fact-store, pipeline)
Rename FactStore.ts->fact-store.ts and SemanticPipeline.ts->pipeline.ts (symbols
unchanged); add @dxos/util dep for trim in nl-to-query. Completes the in-progress
kebab normalization; repairs a corrupted index.ts re-export line.
