+++ packages/core/compute/pipeline-rdf/src/fact-store.test.ts
@@ -7,7 +7,7 @@ import { describe, it } from '@effect/vitest';
 import * as Effect from 'effect/Effect';
 import * as Layer from 'effect/Layer';
 
-import { FactStore } from './FactStore';
+import { FactStore } from './fact-store';
 import { type Fact } from './types';
 
 const mk = (over: Partial<Fact> & Pick<Fact, 'id'>): Fact => ({
