+++ packages/core/compute/pipeline-rdf/src/index.ts
@@ -4,11 +4,12 @@
 
 export * as Type from './types';
 export * from './fact-graph';
-export * from './SemanticPipeline';
+export * from './pipeline';
 export * from './stages';
-export * from './FactStore';
+export * from './fact-store';
 export * from './errors';
 export * from './nl-to-query';
+
 export { buildSparql } from './internal/sparql/query-builder';
 export { normalizePredicate } from './internal/sparql/normalize-predicate';
 export { parseSparqlToQuery } from './internal/sparql/sparql-to-query';
