+++ packages/core/compute/pipeline-rdf/src/pipeline.ts
@@ -9,7 +9,7 @@ import { type AiService } from '@dxos/ai';
 import { Pipeline } from '@dxos/pipeline';
 
 import { type SemanticIndexError } from './errors';
-import { FactStore } from './FactStore';
+import { FactStore } from './fact-store';
 import { type ExtractDocument, type ExtractOptions } from './internal/stages/extract';
 import { indexFactsStage } from './stages';
 import { type Fact } from './types';
