+++ packages/core/compute/pipeline-rdf/tsconfig.json
@@ -25,6 +25,9 @@
       "path": "../../../common/sql-sqlite"
     },
     {
+      "path": "../../../common/util"
+    },
+    {
       "path": "../../echo/echo"
     },
     {
